--- .env
@@ -1,2 +1,0 @@
-SUPABASE_URL=https://sbypqlojufdwlimyhntp.supabase.co
-SUPABASE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InNieXBxbG9qdWZkd2xpbXlobnRwIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MzEzMTc2NzgsImV4cCI6MjA0Njg5MzY3OH0.ZcjOBI11ple5OvobrsYYUiEnxRnaIALFfj6Qk3rQTpU
